fix: 8 critical/high/medium bugs — PWM baseline, overcurrent protection, EMA reset, double→float, PRIMASK compliance, telemetry refactor
CRITICAL fixes:
- Pwm_Driver: Restore V0.0 baseline (Up counting, PartialRemap, PWM1+PWM2, OCNPolarity_Low, OCNIdleState_Set)
- Inverter_Control: Fix freq ramp convergence — exact equality check → tolerance range |diff|≤step
- Ui_Controller: Add overcurrent protection — Soft_Start_Fault() now called at 5A threshold in SWEEPING/RUNNING

HIGH fixes:
- Esp8266_Driver: Clear RX buffer on re-init to prevent stale frame consumption
- Esp8266_Driver/Key_Driver: Fix bare __enable_irq() → PRIMASK save/restore pattern

MEDIUM fixes:
- Adc_Driver: Add compile-time assertion for HSE=72MHz (DWT coprimality with 100kHz PWM)
- Oled_Driver: Change double→float for Cortex-M3 (no hardware FPU)
- Ui_Controller: Lift EMA display state to module level with Reset_Display_EMA() on state transitions
- App_Network: Refactor telemetry gating — nested return → flag variable pattern

Other:
- main.c: Add DBGMCU IWDG debug freeze for safe Flash downloads
- Target 1.BAT: Add fromelf.exe HEX generation step

+++ Keil_Project/Hardware/Oled_Driver.c
@@ -203,22 +203,22 @@ void Oled_Driver_Show_Signed_Num(uint8_t line, uint8_t column, int32_t num, uint
         Oled_Driver_Show_Char(line, column + i + 1, abs_val / Oled_Int_Pow(10, len - i - 1) % 10 + '0');
 }
 
-void Oled_Driver_Show_Float(uint8_t line, uint8_t column, double num,
+void Oled_Driver_Show_Float(uint8_t line, uint8_t column, float num,
                              uint8_t int_len, uint8_t fract_len)
 {
     uint32_t int_part, fract_part;
     uint32_t pow10 = Oled_Int_Pow(10, fract_len);
     uint8_t  offset = 0;
 
-    if (num < 0) {
+    if (num < 0.0f) {
         Oled_Driver_Show_Char(line, column, '-');
         num = -num;
         offset = 1;
     }
 
-    num += 0.5 / (double)pow10;
+    num += 0.5f / (float)pow10;
     int_part   = (uint32_t)num;
-    fract_part = (uint32_t)((num - int_part) * (double)pow10);
+    fract_part = (uint32_t)((num - (float)int_part) * (float)pow10);
 
     Oled_Driver_Show_Num(line, column + offset, int_part, int_len);
     Oled_Driver_Show_Char(line, column + offset + int_len, ':');
